[Bug 4177788] JSONValue cannot take an unsigned long, static casted to uint32_t.
git-svn-id: svn+ssh://svn.corp.yahoo.com/yahoo/yst/websearch/projects/clusterlib/trunk@272196 73f437aa-624d-0410-a3ad-9285e900ec26

+++ src/gui/src/clusterlib_jsonrpc_adaptor.cc
@@ -414,7 +414,7 @@ MethodAdaptor::getNotifyableAttributesFromKey(
         NameList names;
         if (dynamic_pointer_cast<Root >(notifyableSP)) {
             attributes["# of Applications"] =
-                (m_rootSP->getApplicationNames()).size();
+                static_cast<uint32_t>(m_rootSP->getApplicationNames().size());
             attributes[idApplicationSummary] =
                 getApplicationStatus(childObj["applications"].
                                      get<JSONValue::JSONArray>());
